src/enhancements/quantum_inspired_optimization/quantum_inspired_optimization.py
import numpy as np
from qiskit import QuantumCircuit
import qiskit_aer
from typing import List, Tuple

class QuantumInspiredOptimizer:
    def __init__(self, num_qubits: int, num_shots: int = 1000):
        self.num_qubits = num_qubits
        self.num_shots = num_shots
        self.simulator = qiskit_aer.AerSimulator()

    def create_circuit(self) -> QuantumCircuit:
        circuit = QuantumCircuit(self.num_qubits)
        for i in range(self.num_qubits):
            circuit.h(i)  # Apply Hadamard gate to create superposition
        circuit.measure_all()
        return circuit

    def run_circuit(self, circuit: QuantumCircuit) -> dict:
        job = self.simulator.run(circuit, shots=self.num_shots)
        return job.result().get_counts()

    def interpret_results(self, counts: dict) -> List[Tuple[str, float]]:
        total_shots = sum(counts.values())
        sorted_results = sorted(counts.items(), key=lambda x: x[1], reverse=True)
        return [(state, count / total_shots) for state, count in sorted_results]

    def optimize_memory_task(self, task_complexity: float) -> str:
        circuit = self.create_circuit()
        counts = self.run_circuit(circuit)
        results = self.interpret_results(counts)

        # Select the most probable state as the optimized solution
        optimized_state = results[0][0]

        # Map the binary state to a memory enhancement strategy
        strategy = self.map_state_to_strategy(optimized_state, task_complexity)

        return strategy

    def map_state_to_strategy(self, state: str, task_complexity: float) -> str:
        # Convert binary state to integer
        state_int = int(state, 2)

        # Define strategy mapping based on state and task complexity
        strategies = [
            "Chunking",
            "Mnemonic Devices",
            "Mind Mapping",
            "Spaced Repetition",
            "Method of Loci",
            "Elaborative Rehearsal",
            "Dual Coding",
            "Active Recall"
        ]

        # Use state_int and task_complexity to select a strategy
        strategy_index = (state_int + int(task_complexity * 10)) % len(strategies)
        return strategies[strategy_index]

def quantum_inspired_memory_optimization(task_description: str, task_complexity: float) -> str:
    """
    Optimize memory enhancement strategy using quantum-inspired optimization.

    Args:
    task_description (str): Description of the memory task.
    task_complexity (float): Complexity of the task, ranging from 0 to 1.

    Returns:
    str: Optimized memory enhancement strategy.
    """
    num_qubits = 3  # Adjust based on the complexity of the optimization problem
    optimizer = QuantumInspiredOptimizer(num_qubits)
    optimized_strategy = optimizer.optimize_memory_task(task_complexity)

    return f"For the task '{task_description}' with complexity {task_complexity}, " \
           f"the optimized memory enhancement strategy is: {optimized_strategy}"

# Example usage
if __name__ == "__main__":
    task = "Memorize the periodic table"
    complexity = 0.8
    result = quantum_inspired_memory_optimization(task, complexity)
    print(result)
